working rays and basic sphere and plane intersections

+++ srcs/lib/ft_assign.c
@@ -6,13 +6,13 @@
 /*   By: panger <panger@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/07 17:36:55 by panger            #+#    #+#             */
-/*   Updated: 2024/02/07 17:37:39 by panger           ###   ########.fr       */
+/*   Updated: 2024/02/20 13:00:10 by panger           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-t_pair	assign_pair(float x, float y)
+t_pair	assign_pair(double x, double y)
 {
 	t_pair	ret;
